+++ src/async/index.ts
@@ -3,4 +3,5 @@ export * from "./async-context.js";
 export * from "./async.js";
 export * from "./queue.js";
 export * from "./tasks.js";
+export * from "./timers.js";
 export * from "./watch-for.js";